add controllers and views for adding clothings

+++ code/grails-app/domain/com/itfits/Measurement.groovy
@@ -3,7 +3,7 @@ package com.itfits
 class Measurement {
 
     MeasurementType type
-    int value
+    double value
     MeasurementDegree degree
     UserClothingContribution contribution
